+++ Changelog.md
@@ -2,6 +2,8 @@
 
 ## 1.6.0
 
+- Provide contextual menu to transform routes. In the `New Camel file` menu tranform the right clicked files or all the file in a clicked folder.
+
 ## 1.5.0
 
 - Provide completion for Camel component Maven dependency in pom.xml. Take care to have the correct runtime provider selected in Preferences.